Fix multiple-room chatserver
The issue was that communication actions were present in a
non-spawnClient block -- meaning that the actions were evaluated on the
server instead of the client (hence the weird errors).

The way I've fixed this is to add a placeholder div for the room list,
spawn off the communication threads, and replace these with the
rooms once they've been received from the server.

The session runtime is fine; it's just that I should make the type
system smarter, in order to ensure that communication actions aren't
present in code that's run to generate the initial DOM to be delivered
to the client.

Thanks for letting me know!